Introduce draft index page.
The code is based on a free template and was minimally customized to
better reflect the target website.
A number of files were also removed from the template to keep the
repository size small. Some files or sections of css are likely not
required for the website to function but removing them takes more time
than is available.